src/bots/requirements.txt
certifi==2019.11.28
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.7
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.0.4
idna==2.8
marshmallow==3.5.0
marshmallow-enum==1.5.1
oauthlib==3.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.26.0
requests-oauthlib==1.3.0
schedule==0.6.0
six==1.14.0
sseclient-py==1.7
tweepy==3.8.0
urllib3==1.26.7
Werkzeug==0.16.0
zipp==3.1.0
